# src/services/api-server/generated/swagger.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    ExecuteCommandResponse:
      properties:
        success:
          type: boolean
        taskId:
          type: string
        message:
          type: string
      required:
        - success
        - taskId
        - message
      type: object
      additionalProperties: false
    ExecuteCommandRequest:
      properties:
        command:
          type: string
        cwd:
          type: string
      required:
        - command
      type: object
      additionalProperties: false
...
    Record_string.any_:
      properties: {}
      additionalProperties: {}
      type: object
      description: Construct a type with a set of properties K of type T
...
    CommandListItem:
      properties:
        name:
          type: string
        source:
          type: string
        description:
          type: string
        argumentHint:
          type: string
        filePath:
          type: string
        content:
          type: string
      required:
        - name
        - source
        - filePath
      type: object
      additionalProperties: false
    CommandsListResponse:
      properties:
        commands:
          items:
            $ref: "#/components/schemas/CommandListItem"
          type: array
        total:
          type: number
          format: double
      required:
        - commands
        - total
      type: object
      additionalProperties: false
    CommandNamesResponse:
      properties:
        names:
          items:
            type: string
          type: array
        total:
          type: number
          format: double
      required:
        - names
        - total
      type: object
      additionalProperties: false
    CommandDetails:
      properties:
        name:
          type: string
        source:
          type: string
        description:
          type: string
        argumentHint:
          type: string
        filePath:
          type: string
        content:
          type: string
      required:
        - name
        - source
        - filePath
        - content
      type: object
      additionalProperties: false
    CreateCommandResponse:
      properties:
        success:
          type: boolean
        command:
          $ref: "#/components/schemas/CommandDetails"
      required:
        - success
        - command
      type: object
      additionalProperties: false
    CreateCommandRequest:
      properties:
        name:
          type: string
        content:
          type: string
        description:
          type: string
        argumentHint:
          type: string
      required:
        - name
        - content
      type: object
      additionalProperties: false
    UpdateCommandResponse:
      properties:
        success:
          type: boolean
        command:
          $ref: "#/components/schemas/CommandDetails"
      required:
        - success
        - command
      type: object
      additionalProperties: false
    UpdateCommandRequest:
      properties:
        content:
          type: string
        description:
          type: string
        argumentHint:
          type: string
      type: object
      additionalProperties: false
    DeleteCommandResponse:
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
      type: object
      additionalProperties: false
    ExecuteSlashCommandResponse:
      properties:
        success:
          type: boolean
        taskId:
          type: string
        commandName:
          type: string
        commandSource:
          type: string
      required:
        - success
        - taskId
        - commandName
        - commandSource
      type: object
      additionalProperties: false
    ExecuteSlashCommandRequest:
      properties:
        args:
          type: string
        mode:
          type: string
        images:
          items:
            type: string
          type: array
      type: object
      additionalProperties: false
    ValidateCommandResponse:
      properties:
        valid:
          type: boolean
        frontmatter:
          $ref: "#/components/schemas/Record_string.any_"
        contentPreview:
          type: string
        contentLength:
          type: number
          format: double
      required:
        - valid
        - frontmatter
        - contentPreview
        - contentLength
      type: object
      additionalProperties: false
    ValidateCommandRequest:
      properties:
        content:
          type: string
      required:
        - content
      type: object
      additionalProperties: false
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
info:
  title: Hoody Code API
  version: 1.0.0
  description: REST API for Hoody Code VSCode Extension - Remote control and
    integration interface
  license:
    name: Apache-2.0
  contact:
    name: Hoody Code
    url: https://hoody.com
paths:
...
  /terminal/execute:
    post:
      operationId: ExecuteCommand
      responses:
        "200":
          description: Command execution initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteCommandResponse"
      description: Execute a terminal command
      summary: Execute terminal command
      security:
        - bearer: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCommandRequest"
...
  /commands:
    get:
      operationId: ListCommands
      responses:
        "200":
          description: Commands listed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandsListResponse"
      description: List all available slash commands from all sources
      summary: List all commands
      security:
        - bearer: []
      parameters:
        - description: Filter by source (built-in | global | project)
          in: query
          name: source
          required: false
          schema:
            type: string
        - description: "Include full command content (default: false)"
          in: query
          name: includeContent
          required: false
          schema:
            type: string
    post:
      operationId: CreateCommand
      responses:
        "201":
          description: Command created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommandResponse"
      description: Create or update a project-level slash command
      summary: Create command
      security:
        - bearer: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandRequest"
  /commands/names:
    get:
      operationId: GetCommandNames
      responses:
        "200":
          description: Command names retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandNamesResponse"
      description: Get list of command names for autocomplete
      summary: Get command names
      security:
        - bearer: []
      parameters: []
  /commands/{name}:
    get:
      operationId: GetCommand
      responses:
        "200":
          description: Command retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandDetails"
      description: Get a specific slash command by name
      summary: Get command by name
      security:
        - bearer: []
      parameters:
        - description: Command name
          in: path
          name: name
          required: true
          schema:
            type: string
    put:
      operationId: UpdateCommand
      responses:
        "200":
          description: Command updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCommandResponse"
      description: Update an existing project-level slash command
      summary: Update command
      security:
        - bearer: []
      parameters:
        - description: Command name
          in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommandRequest"
    delete:
      operationId: DeleteCommand
      responses:
        "200":
          description: Command deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCommandResponse"
      description: Delete a project-level slash command
      summary: Delete command
      security:
        - bearer: []
      parameters:
        - description: Command name
          in: path
          name: name
          required: true
          schema:
            type: string
  /commands/{name}/execute:
    post:
      operationId: ExecuteCommand
      responses:
        "201":
          description: Command executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteSlashCommandResponse"
      description: Execute a slash command by creating a task
      summary: Execute command
      security:
        - bearer: []
      parameters:
        - description: Command name
          in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteSlashCommandRequest"
  /commands/{name}/validate:
    post:
      operationId: ValidateCommand
      responses:
        "200":
          description: Command validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateCommandResponse"
      description: Validate command content without executing
      summary: Validate command
      security:
        - bearer: []
      parameters:
        - description: Command name
          in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCommandRequest"
...
servers:
  - url: /api/v1/agent
...